requirements.txt
lmdb
pyclipper~=1.3.0.post3
shapely~=2.0.1
addict
matplotlib~=3.6.3
addict
numpy~=1.23.0
tqdm>=4.64.1
opencv-python-headless>=3.4.18.65
rapidfuzz>=2.13.7
Pillow>=9.1.1
PyYAML>=6.0
xml-python>=0.4.3
packaging~=23.0
Cython
lanms>=1.0.2

langchain~=0.0.224
requests~=2.28.1
pydantic~=1.10.11
django~=4.2.3
mindspore~=1.10.1
opencv-python~=4.7.0.72
scipy~=1.9.1
yaml~=0.2.5
djangorestframework~=3.14.0